+++ uninstall.php
@@ -8,8 +8,6 @@
  * @package Vulnz_Agent
  */
 
-declare(strict_types=1);
-
 // If uninstall not called from WordPress, exit.
 if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 	exit;